+++ fmt/fmt_fill.c
@@ -19,5 +19,5 @@ size_t fmt_fill(char* dest,size_t srclen,size_t padlen,size_t maxlen) {
     if (dest>max) break;
     *dest=' '; ++dest;
   }
-  return dest-olddest;
+  return (size_t)(dest-olddest);
 }
