+++ fmt/fmt_long.c
@@ -3,7 +3,7 @@
 size_t fmt_long(char *dest,long int i) {
   if (i<0) {
     if (dest) *dest++='-';
-    return fmt_ulong(dest,-i)+1;
+    return fmt_ulong(dest,(unsigned long)-i)+1;
   } else
-    return fmt_ulong(dest,i);
+    return fmt_ulong(dest,(unsigned long)i);
 }
